requirements.txt
# KB Service Requirements
fastapi==0.109.0
uvicorn==0.27.0
chromadb==0.4.22
requests==2.31.0
pydantic==2.5.3
pydantic-settings==2.1.0
GitPython==3.1.41
pyyaml==6.0.1
aiohttp==3.9.1
python-dotenv==1.0.0
